requirements.txt
numpy==2.3.1
opencv_python==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
streamlit==1.37.1
tensorflow==2.16.1
tensorflow_intel==2.16.1
